Add files via upload
Plant disease detection can be done by looking for a spot on the diseased plant's leaves. The goal of this paper is to create a Disease Recognition Model that is supported by leaf image classification. To detect plant diseases, we are utilizing image processing with a Convolution neural network (CNN). A convolutional neural network (CNN) is a form of artificial neural network that is specifically intended to process pixel input and is used
in image recognition.